feat: startingq hooking the fuzzer
